Remove Deprecated Config Entry
Namely `consent.gics.fhir.defaultPageSize`, which was deprecated in 5.2.
Also switch to HttpClientConfig, and remove the whole
`consent.gics.fhir` specific configuration class.
