+++ trust-center-agent/src/test/resources/application.yaml
@@ -2,8 +2,6 @@ consent:
   gics:
     fhir:
       baseUrl: http://gics:8080/ttp-fhir/fhir/gics
-      auth:
-        none: { }
     pageSize: 200
 
 deIdentification:
