+++ util/src/main/java/care/smith/fts/util/HttpClientConfig.java
@@ -6,10 +6,12 @@ import static java.util.Objects.requireNonNull;
 import care.smith.fts.util.auth.HttpClientAuth;
 import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotBlank;
+import org.springframework.boot.context.properties.bind.ConstructorBinding;
 
 public record HttpClientConfig(
     @NotBlank String baseUrl, @Nullable HttpClientAuth.Config auth, @Nullable Ssl ssl) {
 
+  @ConstructorBinding
   public HttpClientConfig(@NotBlank String baseUrl, HttpClientAuth.Config auth, Ssl ssl) {
     this.baseUrl = requireNonNull(emptyToNull(baseUrl), "Base URL must not be null");
     this.auth = auth;
